+++ server/server.py
@@ -1,90 +1,96 @@
 # Written by Austin Shouli
-
+import sys
+import os
 from flask import Flask, render_template, redirect, url_for, request, flash
 
-app = Flask(__name__)
+from data.authentication.user_authentication import user_login
+from data.data_controller.account_data_controller import write_account
+from data.data_controller.event_data_controller import write_event
+from src.account_controller import AccountController, read_user_account
+from src.accounts import Account
+from src.events import Event
+
+app = Flask(__name__, root_path=os.path.dirname(__file__), static_folder='server/static',
+            template_folder='server/templates')
+
 app.secret_key = 'super secret key'
 
+# from data import tables
 
-from src import accounts
-from src import events
-from data.data_controller import account_data_controller
-from data.data_controller import event_data_controller
-from data.authentication import user_authentication
+# tables.create_tables()
 
-from data import tables
 
-#tables.create_tables()
+account_controller = AccountController()
 
-import sys
 
-AcController = accounts.AccountController()
-
-#renders a super simple calendar
+# renders a super simple calendar
 @app.route('/calendar')
 def show_calendar():
     return render_template('calendar.html')
-    
+
+
 @app.route('/event_input', methods=["POST", "GET"])
-def event_input(): 
+def event_input():
     if request.method == "POST":
         user = request.form["nm"]
-        
+
         return redirect(url_for("event", usr=user))
     else:
         return render_template('event_input.html')
 
-#clicking on view calendar in navbar will render a calendar
+
+# clicking on view calendar in navbar will render a calendar
 @app.route('/home')
 def home_view():
     return render_template('base.html')
+
 
 @app.route("/event_input/<usr>")
 def event(usr):
     return f"<h1>Post successful {usr}!</h1>"
 
-#loads login page on application launch
+
+# loads login page on application launch
 @app.route('/')
 def index():
     return render_template('index.html')
 
 
-#handles submission of login form
+# handles submission of login form
 @app.route('/login', methods=('GET', 'POST'))
 def login():
     if request.method == 'POST':
 
-        loginAccount = accounts.Account()
+        login_account = Account()
 
-        loginAccount.username = request.form['username']
-        loginAccount.password = request.form['password']
+        login_account.username = request.form['username']
+        login_account.password = request.form['password']
 
-        loginAtempt = user_authentication.login(loginAccount)
+        login_attempt = user_login(login_account)
 
-        if loginAtempt[0] == True: 
+        if login_attempt[0]:
 
+            # set activeUser's attributes to the database values
+            account_controller.active_user = read_user_account(login_attempt[1])
 
-            #set activeUser's attributes to the database values
-            AcController.active_user = read_account()
+            print("Account logged in is: {}".format(account_controller.active_user), file=sys.stdout)
+            print("Account Type is: {}".format(account_controller.active_user.account_type), file=sys.stdout)
 
-            print("Account logged in is: {}".format(AcController.active_user), file=sys.stdout)
-            print("Account Type is: {}".format(AcController.active_user.account_type), file=sys.stdout)
-
-            #testing get all subs
-            # allSubs = AcController.getSubscriptions()
+            # testing get all subs
+            # allSubs = account_controller.getSubscriptions()
 
             # print("All Subs are: {}".format(*allSubs), file=sys.stdout)
 
-            #end testing
+            # end testing
 
-            print("Server Sees Logged In. accountID is: {}".format(loginAtempt[1]), file=sys.stdout)
+            print("Server Sees Logged In. accountID is: {}".format(login_attempt[1]), file=sys.stdout)
             return render_template('base.html')
 
-        else: 
-  
-            print("Database Exception: {}", loginAtempt[1], file=sys.stdout) #Debugging: sql error to terminal
+        else:
+
+            print("Database Exception: {}", login_attempt[1], file=sys.stdout)  # Debugging: sql error to terminal
             flash('Your Account was not found. Please try again.')
-            return redirect(url_for('index')) 
+            return redirect(url_for('index'))
 
     else:
         print("Test Point 1", sys.stdout)
@@ -92,91 +98,86 @@ def login():
         return redirect(url_for('success1', name=user))
 
 
-#handles create account form submission
+# handles create account form submission
 @app.route('/create_account', methods=['GET', 'POST'])
 def create_account():
     if request.method == 'POST':
 
-        #try to create account in DB. 
-        newAccount = accounts.Account()
-        newAccount.username = request.form['username']
-        newAccount.password = request.form['password']
-        newAccount.account_type = request.form['type']
-        newAccount.first_name = request.form['fname']
-        newAccount.last_name = request.form['lname']
-        newAccount.email = request.form['email']
+        # try to create account in DB.
+        new_account = Account()
+        new_account.username = request.form['username']
+        new_account.password = request.form['password']
+        new_account.account_type = request.form['type']
+        new_account.first_name = request.form['first_name']
+        new_account.last_name = request.form['last_name']
+        new_account.email = request.form['email']
 
-        result = account_data_controller.write_account(newAccount)
+        result = write_account(new_account)
 
-        #debugging
-        #print("Result of write is: " + {result[0]} + result[1])
-        
-        if result[0] == True:
-            flash('Account Succesfully created. Please log in')
+        # debugging
+        # print("Result of write is: " + {result[0]} + result[1])
 
-        else: 
+        if result[0]:
+            flash('Account successfully created. Please log in')
 
-            print("DB Error: {}", result[1], file=sys.stdout) #Debugging: sql error to terminal
+        else:
+
+            print("DB Error: {}", result[1], file=sys.stdout)  # Debugging: sql error to terminal
             flash('There was an error creating your account. Please try again.')
-        
+
         return redirect(url_for('index'))
 
     return render_template('create_account.html')
 
 
-#handles log out button
+# handles log out button
 @app.route('/logout', methods=['GET', 'POST'])
 def sign_out():
-     
-     AcController.active_user = None
+    account_controller.active_user = None
 
-     return redirect(url_for('index'))
+    return redirect(url_for('index'))
 
 
-#handles navigation to Create Event page
+# handles navigation to Create Event page
 @app.route('/home', methods=['GET', 'POST'])
 def returnHome():
-
     return render_template('base.html')
-    
 
-#handles navigation to the home page
+
+# handles navigation to the home page
 @app.route('/NavCreateEvent', methods=['GET', 'POST'])
 def NavCreateEvent():
-
     return render_template('create_event.html')
-    
 
-#handles submission of create event form 
+
+# handles submission of create event form
 @app.route('/createEvent', methods=['GET', 'POST'])
 def createEvent():
     if request.method == 'POST':
 
-        #handle submit for create event
-        newEvent = events.Event()
+        # handle submit for create event
+        new_event = Event()
 
-        newEvent.name = request.form['event_name']
-        newEvent.category = request.form['category']
-        newEvent.start_date = request.form['start_date']
-        newEvent.end_date = request.form['end_date']
-        newEvent.visibility = request.form['visibility']
+        new_event.name = request.form['event_name']
+        new_event.category = request.form['category']
+        new_event.start_date = request.form['start_date']
+        new_event.end_date = request.form['end_date']
+        new_event.visibility = request.form['visibility']
 
-        result = event_data_controller.write_event(newEvent)        
+        result = write_event(new_event)
 
-        if result[0] == True:
-            flash('Event Succesfully created')
-            print("DB Write Event Successful", file=sys.stdout) #Debugging: sql error to terminal
-        
-        else: 
+        if result[0]:
+            flash('Event successfully created')
+            print("DB Write Event Successful", file=sys.stdout)  # Debugging: sql error to terminal
 
-            print("DB Error: {}", result[1], file=sys.stdout) #Debugging: sql error to terminal
+        else:
+
+            print("DB Error: {}", result[1], file=sys.stdout)  # Debugging: sql error to terminal
             flash('There was an error creating your event. Please try again.')
-        
 
         return render_template('base.html')
 
 
-#sets host port and debug mode
+# sets host port and debug mode
 if __name__ == '__main__':
-    app.run(host="localhost", port=3000, debug=True)
-
+    app.run(host="localhost", port=1234, debug=True)
