+++ server/static/CSS/main.css
@@ -1,21 +1,17 @@
 /* Written by Austin Shouli */
 
 
-
-.login  {
+.login {
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
     height: 200px;
 
-    margin-top: 20%;
-
     width: 50%;
-    margin: auto; 
+    margin: auto;
 
-    
-    
+
     border: 3px solid black;
     padding: 10px;
 
@@ -29,7 +25,6 @@
     justify-content: right;
 
 
-    
 }
