+++ server/templates/base.html
@@ -1,61 +1,70 @@
 <!doctype html>
 <!--For bootstrap intro: https://getbootstrap.com/docs/4.3/getting-started/introduction/-->
-<html>
-    <head>
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-        <!-- <link rel="stylesheet" href="../static/CSS/calendar.css"> -->
-        <title>    
-            {% block title %}
-            {% endblock %} 
-            
-        </title>
-        
-    </head>
-    
-    <body>
-        <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-            <a style="padding-left: 10px" class="navbar-brand" href="/home">Raven</a>
-            <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-              <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-              <ul class="navbar-nav">
-                <li class="nav-item active">
-                  <a class="nav-link" href="/home">Home <span class="sr-only">(current)</span></a>
-                </li>
-                <li class="nav-item">
-                  <a class="nav-link" href="/NavCreateEvent">Create Event</a>
-                </li>
-                <li class="nav-item">
-                  <a class="nav-link" href="calendar">View Calendar</a>
-                </li>
-                <li class="nav-item">
-                  <a class="nav-link" href="#">Help</a>
-                </li>
-                <li class="nav-item">
-                  <a class="nav-link disabled" href="#" tabindex="-1" aria-disabled="true">Prioritize Events</a>
-                <li class="nav-item">
-                  <a class="nav-link" href="/logout" tabindex="-1" aria-disabled="true">Sign Out</a>
-                </li>
-              </ul>
-            </div>
-          </nav>
-        <h1>{% block test %}{% endblock %}</h1>
-        {% block content %}{% endblock %}
-        
-        <div class="container" style="width: 30%; position:fixed; left: 50%; padding-top: 40px;; transform: translate(-50%, 0);">
-          {% with messages = get_flashed_messages() %}
-          {%  if  messages %}
-          {%  for  message  in  messages %} 
-          <h3 style="text-align: center;">{{ message }}</h3>
-          {% endfor %}
-          {% endif %}
-          {% endwith %}
-      </div>
-        
-        <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/popper.js@1.14.7/dist/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
-    </body>
-    
+<html lang="en">
+<head>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
+    <!-- <link rel="stylesheet" href="../static/CSS/calendar.css"> -->
+    <title>
+        {% block title %}
+        {% end-block %}
+
+    </title>
+
+</head>
+
+<body>
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+    <a style="padding-left: 10px" class="navbar-brand" href="/home">Raven</a>
+    <button class="navbar-toggle" type="button" data-toggle="collapse" data-target="#navbarNav"
+            aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggle-icon"></span>
+    </button>
+    <div class="collapse navbar-collapse" id="navbarNav">
+        <ul class="navbar-nav">
+            <li class="nav-item active">
+                <a class="nav-link" href="/home">Home <span class="sr-only">(current)</span></a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/NavCreateEvent">Create Event</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="/calendar">View Calendar</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="#">Help</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link disabled" href="#" tabindex="-1" aria-disabled="true">Prioritize Events</a>
+            <li class="nav-item">
+                <a class="nav-link" href="/logout" tabindex="-1" aria-disabled="true">Sign Out</a>
+            </li>
+        </ul>
+    </div>
+</nav>
+<h1>{% block test %}{% end-block %}</h1>
+{% block content %}{% end-block %}
+
+<div class="container"
+     style="width: 30%; position:fixed; left: 50%; padding-top: 40px;; transform: translate(-50%, 0);">
+    {% with messages = get_flashed_messages() %}
+    {% if messages %}
+    {% for message in messages %}
+    <h3 style="text-align: center;">{{ message }}</h3>
+    {% end-for %}
+    {% endif %}
+    {% end-with %}
+</div>
+
+<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+        integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+        crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/popper.js@1.14.7/dist/umd/popper.min.js"
+        integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+        crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/js/bootstrap.min.js"
+        integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+        crossorigin="anonymous"></script>
+</body>
+
 </html>
