+++ server/templates/create_account.html
@@ -1,79 +1,95 @@
- <!-- Written by Austin Shouli -->
-
+<!-- Written by Austin Shouli -->
 
 
 <!doctype html>
 
 
 <head>
-    <title>Create Account</title> 
+    <title>Create Account</title>
 
     <meta charset=utf-8>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
     <link rel="stylesheet" href="{{ url_for('static', filename='../css/main.css') }}">
 
-    </head>
-   
-    <body>
-        
+</head>
+
+<body>
+
+
+<div class="container">
+    <h1>Create Account</h1>
+
+
+    <form name="Create" action="{{ url_for('create_account') }}" method="post" class="login_fields">
+
 
         <div class="container">
-            <h1>Create Account</h1>
-        
-
-            <form name="Create" action="{{ url_for('create_account') }}" method="post" class="login_fields">
-    
-               
-                <div class="container">
-                    <label>Username: </label> 
-                    <input type="text" name="username">
-                </div>
-                <br>
-               
-                <div class="container">
-                    <label>Password: </label>
-                    <input type="text" name="password">
-                </div>
-                <br>
-
-                <div class="container">
-                    <label>Type: </label>
-                    <select name="type" id="type">
-                        <option value="1">Student</option>
-                        <option value="2">Professor</option>
-                    </select>
-                </div>
-                <br>
-
-                <div class="container">
-                    <label>First Name: </label>
-                    <input type="text" name="fname">  
-                </div>
-                <br>
-
-                <div class="container">
-                    <label>Last Name: </label>
-                    <input type="text" name="lname">
-                </div>
-                <br>
-               
-                <div class="container">
-                    <label>Email: </label> 
-                    <input type="text" name="email">
-                </div>
-                <br>
-    
-                <div class="container">
-                    <button type="submit" class="btn btn-primary" value="Submit">Submit</button>
-                    <button type="reset" class="btn btn-danger" value="Reset">Reset</button>
-                </div>
-                <br>
-
-            </form>
+            <label>Username: </label>
+            <label>
+                <input type="text" name="username">
+            </label>
         </div>
+        <br>
+
+        <div class="container">
+            <label>Password: </label>
+            <label>
+                <input type="text" name="password">
+            </label>
+        </div>
+        <br>
+
+        <div class="container">
+            <label>Type: </label>
+            <label for="type"></label><select name="type" id="type">
+            <option value="1">Student</option>
+            <option value="2">Professor</option>
+        </select>
+        </div>
+        <br>
+
+        <div class="container">
+            <label>First Name: </label>
+            <label>
+                <input type="text" name="firstname">
+            </label>
+        </div>
+        <br>
+
+        <div class="container">
+            <label>Last Name: </label>
+            <label>
+                <input type="text" name="lastname">
+            </label>
+        </div>
+        <br>
+
+        <div class="container">
+            <label>Email: </label>
+            <label>
+                <input type="text" name="email">
+            </label>
+        </div>
+        <br>
+
+        <div class="container">
+            <button type="submit" class="btn btn-primary" value="Submit">Submit</button>
+            <button type="reset" class="btn btn-danger" value="Reset">Reset</button>
+        </div>
+        <br>
+
+    </form>
+</div>
 
 
-        <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/popper.js@1.14.7/dist/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
-    </body>
+<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+        integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+        crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/popper.js@1.14.7/dist/umd/popper.min.js"
+        integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+        crossorigin="anonymous"></script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.3.1/dist/js/bootstrap.min.js"
+        integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+        crossorigin="anonymous"></script>
+</body>
