+++ server/templates/create_category.html
@@ -5,80 +5,85 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Event</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
 
 </head>
 <body>
-    <!--Add names to href when done or just flask inherit base.html-->
-    <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-        <a class="navbar-brand" href="#">Raven</a>
-        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-          <span class="navbar-toggler-icon"></span>
-        </button>
-        <div class="collapse navbar-collapse" id="navbarNav">
-          <ul class="navbar-nav">
+<!--Add names to href when done or just flask inherit base.html-->
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+    <a class="navbar-brand" href="#">Raven</a>
+    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav"
+            aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+    </button>
+    <div class="collapse navbar-collapse" id="navbarNav">
+        <ul class="navbar-nav">
             <li class="nav-item active">
-              <a class="nav-link" href="#">Home</a>
+                <a class="nav-link" href="#">Home</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="#">Create Event</a>
+                <a class="nav-link" href="#">Create Event</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="#">View Calendar</a>
+                <a class="nav-link" href="#">View Calendar</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="#">Help</a>
+                <a class="nav-link" href="#">Help</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link disabled" href="#" tabindex="-1" aria-disabled="true">Prioritize Events</a>
+                <a class="nav-link disabled" href="#" tabindex="-1" aria-disabled="true">Prioritize Events</a>
             </li>
-          </ul>
-        </div>
-    </nav>
+        </ul>
+    </div>
+</nav>
 
-    <!--Create event form-->
+<!--Create event form-->
+<br>
+<form>
+    <div class="container">
+        <label for="Category">Category Name:</label>
+        <input type="text" class="form-control" id="Category" placeholder="Enter category name">
+    </div>
     <br>
-    <form>
-        <div class="container">
-          <label for="Category">Category Name:</label>
-          <input type="text" class="form-control" id="Category" placeholder="Enter category name"> 
-        </div>
-        <br>
-        <div class="container">
-          <label for="Date">Date of Category:</label>
-          <input type="datetime-local" class="form-control" id="Date">
-          <!--change above to just "date if time not needed"-->
-        </div>
-        <br>
-        <div class="container">
-          <label for="Section">Section #:</label>
-          <input type="number" class="form-control" id="Section">
-        </div>
-        <br>
-        <div class="container">
-            <label for="info">Any additional Info:</label>
-            <textarea class="form-control" id="info" rows="3"></textarea>
-        </div>
-        <br>
-        <div class="container">
+    <div class="container">
+        <label for="Date">Date of Category:</label>
+        <input type="datetime-local" class="form-control" id="Date">
+        <!--change above to just "date if time not needed"-->
+    </div>
+    <br>
+    <div class="container">
+        <label for="Section">Section #:</label>
+        <input type="number" class="form-control" id="Section">
+    </div>
+    <br>
+    <div class="container">
+        <label for="info">Any additional Info:</label>
+        <textarea class="form-control" id="info" rows="3"></textarea>
+    </div>
+    <br>
+    <div class="container">
+        <label>
             <select class="custom-select">
                 <option selected>Select the type</option>
                 <option value="1">One</option>
                 <option value="2">Two</option>
                 <option value="3">Three</option>
             </select>
-        </div>
-        <br>
-        <div class="container">
-          <input class="form-check-input" type="checkbox" value="" id="invalidCheck">
-          <label class="form-check-label" for="invalidCheck">
-          Make it public?
-        </div>
-        <br>
-        <div class="container">
-          <button type="submit" class="btn btn-primary">Submit</button>
-          <button type="reset" class="btn btn-danger">Reset</button>
-        </div>
-      </form>
+        </label>
+    </div>
+    <br>
+    <div class="container">
+        <input class="form-check-input" type="checkbox" value="" id="invalidCheck">
+        <label class="form-check-label" for="invalidCheck">
+            Make it public?
+        </label>
+    </div>
+    <br>
+    <div class="container">
+        <button type="submit" class="btn btn-primary">Submit</button>
+        <button type="reset" class="btn btn-danger">Reset</button>
+    </div>
+</form>
 </body>
 </html>
