+++ server/templates/event_input.html
@@ -9,8 +9,10 @@
 {% block content %}
     <form action="#" method="post">
         <p>Name:</p>
-        <p><input type="text" name="nm"/></p>
+        <p><label>
+            <input type="text" name="nm"/>
+        </label></p>
         <p><input type="submit" value="submit"/></p>
-    
+
     </form>
 {% endblock %}
