+++ server/templates/index.html
@@ -1,70 +1,74 @@
 <!-- Written by Austin Shouli -->
 
 
-
 <!doctype html>
-   
+
 
 <head>
-    
-    <title>Login</title> 
-    <meta charset=utf-8> 
 
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
+    <title>Login</title>
+    <meta charset=utf-8>
+
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
 
     <link rel="stylesheet" href="{{ url_for('static', filename='css/main.css') }}">
 
 </head>
-    <body>
+<body>
 
-       <div class="">
-        
-
-           <div class="container" style="width: 30%; padding-top: 30px;">
-                <br>
-                <h1 style="margin: 30 px; text-align: center;">RAVEN</h1>
-                <h3 style="margin: 30 px; text-align: center;">Calendar Application</h3>
-                
-                <div class="container">
-                    <!-- Snippet to flash messages -->
-                    {% with messages = get_flashed_messages() %}
-                    {%  if  messages %}
-                    {%  for  message  in  messages %} 
-                    <p style="text-align: center;">{{ message }}</p>
-                    {% endfor %}
-                    {% endif %}
-                    {% endwith %}
-                </div>
-                
-                    
-                <!-- <form name="login" class="login_fields" action="http://localhost:3000/login" method="post"> -->
-                <form name="login" class="login_fields" action="{{ url_for('login') }}" method="post">
-                
-                    <div class="container">
-                        <label>Username: </label>
-                        <input class="form-control" type="text" name="username">
-                    </div>
-                    <br>
-                
-                    <div class="container">
-                        <label>Password: </label> 
-                        <input class="form-control" type="password" name="password">
-                    </div>
-                    <br>
-                
-                    <div class="container">
-                        <button type="submit" class="btn btn-primary" value="Submit">Login</button>
-                       
-                        <a class="btn btn-secondary" href="{{ url_for('create_account') }}">Create Account</a>
-                    </div>
-                    <br>
+<div class="">
 
 
-                </form>
-                
-                
-           </div>
-        
+    <div class="container" style="width: 30%; padding-top: 30px;">
+        <br>
+        <h1 style="margin: 30px; text-align: center;">RAVEN</h1>
+        <h3 style="margin: 30px; text-align: center;">Calendar Application</h3>
+
+        <div class="container">
+            <!-- Snippet to flash messages -->
+            {% with messages = get_flashed_messages() %}
+            {% if  messages %}
+            {% for  message  in  messages %}
+            <p style="text-align: center;">{{ message }}</p>
+            {% end-for %}
+            {% endif %}
+            {% endwith %}
         </div>
 
-    </body>
+
+        <!-- <form name="login" class="login_fields" action="http://localhost:3000/login" method="post"> -->
+        <form name="login" class="login_fields" action="{{ url_for('login') }}" method="post">
+
+            <div class="container">
+                <label>Username: </label>
+                <label>
+                    <input class="form-control" type="text" name="username">
+                </label>
+            </div>
+            <br>
+
+            <div class="container">
+                <label>Password: </label>
+                <label>
+                    <input class="form-control" type="password" name="password">
+                </label>
+            </div>
+            <br>
+
+            <div class="container">
+                <button type="submit" class="btn btn-primary" value="Submit">Login</button>
+
+                <a class="btn btn-secondary" href="{{ url_for('create_account') }}">Create Account</a>
+            </div>
+            <br>
+
+
+        </form>
+
+
+    </div>
+
+</div>
+
+</body>
