+++ server/templates/logged_in.html
@@ -2,5 +2,5 @@
 
 <body>
 
-    <h1>You have logged in!</h1>
+<h1>You have logged in!</h1>
 </body>
